Internal modules are back
the implementation details which are subject to change are in Internal,
the implementation details which implemententers of custom recursion
schemes need to know about and which must thus be stable are in Custom.
