+++ src/RecursionSchemes.hs
@@ -5,18 +5,17 @@ module RecursionSchemes
   , Recur(..), CataPos, cata
   , Untouched(..), ParaPosT, paraT, para
   , Fast(..), ZygoPosT, withZygotizedRecFun, zygoT, zygo
-  , Cofree, histoT, histo
+  , HistoPosT, histoT, histo
 
   , CorecFun, runCorecFun, corecFun
   , Seed(..), AnaPos, ana
   , Done(..), ApoPosT, apoT, apo
   , Switch(..), GapoPosT, gapoT, gapo
-  , Free, futuT, futu
+  , FutuPosT, futuT, futu
   ) where
-
-import Control.Comonad.Cofree
-import Control.Monad.Free
 
 import Data.Functor.Foldable hiding (cata, para, zygo, histo, ana, apo, gapo, futu)
 import RecursionSchemes.Custom.Gather
 import RecursionSchemes.Custom.Scatter
+import RecursionSchemes.Internal.Gather
+import RecursionSchemes.Internal.Scatter
